Only show included concepts
