+++ src/heliokos/ui/templates/scheme.html
@@ -21,10 +21,11 @@
     </div>
     <div class="tablet:grid-col-6 margin-bottom-3">
 
-        {% if concepts | length > 0 %}
+    {% if concepts | length > 0 %}
+
         <h2>Add a concept to this scheme</h2>
 
-        <hk-form prevent-default msg-submitting='Adding the "${selected_concept}" concept...' msg-success='The "${selected_concept}" concept was added to the scheme.' target='[data-hk-list="scheme-concepts"]'>
+        <hk-form prevent-default msg-submitting='Adding the "${selected_concept}" concept...' msg-success='The "${selected_concept}" concept was added to the scheme.' target='[data-hk-list="scheme-concepts"], [data-hk-section="define-relationship"]'>
             <form class="usa-form" method="post" action="/conceptscheme/{{scheme.id_suffix}}/concept">
                 <label class="usa-label" for="selected_concept">Pick a Concept</label>
                 <select class="usa-select" id="selected_concept" name="selected_concept" required>
@@ -39,47 +40,61 @@
         </hk-form>
 
 
-        <h2 class="padding-top-3">Define relationships between concepts</h2>
+        <div data-hk-section="define-relationship">
 
-        <hk-form prevent-default msg-submitting='Defining a new scheme concept relationship...' msg-success='The concept relationship was added to the scheme.'>
-            <form class="usa-form" method="post" action="/conceptscheme/{{scheme.id_suffix}}/relation">
-                <label class="usa-label" for="subject_concept_id">Pick a Concept</label>
-                <select class="usa-select" id="subject_concept_id" name="subject_concept_id" required>
-                    <option value>- Select -</option>
-                    {% for concept in concepts %}
-                    <option value="{{concept.id_suffix}}">{{concept.pref_label}}</option>
-                    {% endfor %}
-                </select>
+            <h2 class="padding-top-3">Define relationships between concepts</h2>
 
-                <fieldset class="usa-fieldset margin-top-3">
-                    <legend class="usa-legend">Define how it's related</legend>
-                    <div class="usa-radio">
-                        <input class="usa-radio__input" id="relation_narrower" type="radio" name="relation" value="narrower" checked required>
-                        <label class="usa-radio__label" for="relation_narrower">is a broader concept than (<code>skos:narrower</code>)</label>
-                    </div>
-                    <div class="usa-radio">
-                        <input class="usa-radio__input" id="relation_related" type="radio" name="relation" value="related" required>
-                        <label class="usa-radio__label" for="relation_related">is related to (<code>skos:related</code>)</label>
-                    </div>
-                </fieldset>
+            {% if scheme.concepts | length > 1 %}
+            <hk-form prevent-default msg-submitting='Defining a new scheme concept relationship...' msg-success='The concept relationship was added to the scheme.'>
+                <form class="usa-form" method="post" action="/conceptscheme/{{scheme.id_suffix}}/relation">
+                    <label class="usa-label" for="subject_concept_id">Pick a Concept</label>
+                    <hk-select disable-match="#object_concept_id">
+                        <select class="usa-select" id="subject_concept_id" name="subject_concept_id" required>
+                            <option value>- Select -</option>
+                            {% for concept in scheme.concepts %}
+                            <option value="{{concept.id_suffix}}">{{concept.pref_label}}</option>
+                            {% endfor %}
+                        </select>
+                    </hk-select>
 
-                <label class="usa-label" for="object_concept_id">Pick another Concept</label>
-                <select class="usa-select" id="object_concept_id" name="object_concept_id" required>
-                    <option value>- Select -</option>
-                    {% for concept in concepts %}
-                    <option value="{{concept.id_suffix}}">{{concept.pref_label}}</option>
-                    {% endfor %}
-                </select>
+                    <fieldset class="usa-fieldset margin-top-3">
+                        <legend class="usa-legend">Define how it's related</legend>
+                        <div class="usa-radio">
+                            <input class="usa-radio__input" id="relation_narrower" type="radio" name="relation" value="narrower" checked required>
+                            <label class="usa-radio__label" for="relation_narrower">is a broader concept than (<code>skos:narrower</code>)</label>
+                        </div>
+                        <div class="usa-radio">
+                            <input class="usa-radio__input" id="relation_related" type="radio" name="relation" value="related" required>
+                            <label class="usa-radio__label" for="relation_related">is related to (<code>skos:related</code>)</label>
+                        </div>
+                    </fieldset>
 
-                <button class="usa-button">Define Concept Relationship</button>
-            </form>
-        </hk-form>
+                    <label class="usa-label" for="object_concept_id">Pick another Concept</label>
+                    <hk-select disable-match="#subject_concept_id">
+                        <select class="usa-select" id="object_concept_id" name="object_concept_id" required>
+                            <option value>- Select -</option>
+                            {% for concept in scheme.concepts %}
+                            <option value="{{concept.id_suffix}}">{{concept.pref_label}}</option>
+                            {% endfor %}
+                        </select>
+                    </hk-select>
 
-        {% else %}
+                    <button class="usa-button">Define Concept Relationship</button>
+                </form>
+            </hk-form>
+            {% else %}
+            <p>Add a few concepts to this scheme first.</p>
+            {% endif %}
+
+        </div>
+
+    {% else %}
+
         <h2>Concepts</h2>
         <p>You don't have any concepts yet.</p>
         <p><a class="usa-button" href="/concept">Create a Concept</a></p>
-        {% endif %}
+
+    {% endif %}
 
     </div>
 </div>
